requirements.txt
flask>=3.0
python-dotenv>=1.0
requests>=2.31

# AI providers — install the ones you need:
openai>=1.0
anthropic>=0.30

# For Ollama: uses the openai package with a custom base_url, no extra install needed.
